Add options to allow extracting bones within a bone collection
Adds an extract mode option to the deform rig generator that allows the user to specify a bone collection containing bones to be extracted for the game rig. "Collection" mode extracts only bones within the collection, while "Deform and Collection" mode extracts deform bones as well as bones within the collection. This is useful for generating rigs that require the inclusion of non-deform bones for the purposes of in-game IK or procedural animation.